refactor(project): refactor AccountStore with Zustand
