+++ src/components/Account/Account.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 
 import customer from '../../fixtures/customer.json';
-import { AccountStore } from '../../stores/AccountStore';
+import { useAccountStore } from '../../stores/AccountStore';
 
 import Account from './Account';
 
@@ -10,9 +10,9 @@ import type { Consent } from '#types/account';
 
 describe('<Account>', () => {
   test('renders and matches snapshot', () => {
-    AccountStore.update((s) => {
-      s.user = customer;
-      s.publisherConsents = Array.of({ name: 'marketing', label: 'Receive Marketing Emails' } as Consent);
+    useAccountStore.setState({
+      user: customer,
+      publisherConsents: Array.of({ name: 'marketing', label: 'Receive Marketing Emails' } as Consent),
     });
 
     const { container } = render(<Account panelClassName={'panel-class'} panelHeaderClassName={'header-class'} />);
