+++ src/containers/AccountModal/AccountModal.tsx
@@ -6,7 +6,7 @@ import useQueryParam from '../../hooks/useQueryParam';
 import { addQueryParam, removeQueryParam } from '../../utils/history';
 import PaymentFailed from '../../components/PaymentFailed/PaymentFailed';
 import Welcome from '../../components/Welcome/Welcome';
-import { AccountStore } from '../../stores/AccountStore';
+import { useAccountStore } from '../../stores/AccountStore';
 import LoadingOverlay from '../../components/LoadingOverlay/LoadingOverlay';
 import { ConfigStore } from '../../stores/ConfigStore';
 
@@ -28,7 +28,7 @@ const AccountModal = () => {
   const viewParam = useQueryParam('u');
   const [view, setView] = useState(viewParam);
   const message = useQueryParam('message');
-  const { loading, auth } = AccountStore.useState((s) => s);
+  const { loading, auth } = useAccountStore();
   const config = ConfigStore.useState((s) => s.config);
   const {
     assets: { banner },
