+++ src/stores/CheckoutStore.ts
@@ -5,7 +5,7 @@ import type { CreateOrderPayload, Offer, Order, PaymentMethod, UpdateOrderPayloa
 import { getLocales } from '../services/account.service';
 
 import { ConfigStore } from './ConfigStore';
-import { AccountStore } from './AccountStore';
+import { useAccountStore } from './AccountStore';
 
 type CheckoutStore = {
   offer: Offer | null;
@@ -23,7 +23,7 @@ export const createOrder = async (offerId: string, paymentMethodId?: number) => 
   const {
     config: { cleengId, cleengSandbox },
   } = ConfigStore.getRawState();
-  const { user, auth } = AccountStore.getRawState();
+  const { user, auth } = useAccountStore.getState();
 
   if (!cleengId) throw new Error('cleengId is not configured');
   if (!user || !auth) throw new Error('user is not logged in');
@@ -60,7 +60,7 @@ export const updateOrder = async (orderId: number, paymentMethodId?: number, cou
   const {
     config: { cleengId, cleengSandbox },
   } = ConfigStore.getRawState();
-  const { user, auth } = AccountStore.getRawState();
+  const { user, auth } = useAccountStore.getState();
 
   if (!cleengId) throw new Error('cleengId is not configured');
   if (!user || !auth) throw new Error('user is not logged in');
@@ -93,7 +93,7 @@ export const getPaymentMethods = async () => {
   const {
     config: { cleengId, cleengSandbox },
   } = ConfigStore.getRawState();
-  const { user, auth } = AccountStore.getRawState();
+  const { user, auth } = useAccountStore.getState();
   const { paymentMethods } = CheckoutStore.getRawState();
 
   if (paymentMethods) return paymentMethods; // already fetched payment methods
@@ -115,7 +115,7 @@ export const paymentWithoutDetails = async () => {
   const {
     config: { cleengId, cleengSandbox },
   } = ConfigStore.getRawState();
-  const { user, auth } = AccountStore.getRawState();
+  const { user, auth } = useAccountStore.getState();
   const { order } = CheckoutStore.getRawState();
 
   if (!order) throw new Error('No order created');
@@ -134,14 +134,21 @@ export const adyenPayment = async (paymentMethod: AdyenPaymentMethod) => {
   const {
     config: { cleengId, cleengSandbox },
   } = ConfigStore.getRawState();
-  const { user, auth } = AccountStore.getRawState();
+  const { user, auth } = useAccountStore.getState();
   const { order } = CheckoutStore.getRawState();
 
   if (!order) throw new Error('No order created');
   if (!cleengId) throw new Error('cleengId is not configured');
   if (!user || !auth) throw new Error('user is not logged in');
 
-  const response = await checkoutService.paymentWithAdyen({ orderId: order.id, card: paymentMethod }, cleengSandbox, auth.jwt);
+  const response = await checkoutService.paymentWithAdyen(
+    {
+      orderId: order.id,
+      card: paymentMethod,
+    },
+    cleengSandbox,
+    auth.jwt,
+  );
 
   if (response.errors.length > 0) throw new Error(response.errors[0]);
   if (response.responseData.rejectedReason) throw new Error(response.responseData.rejectedReason);
@@ -153,14 +160,23 @@ export const paypalPayment = async (successUrl: string, cancelUrl: string, error
   const {
     config: { cleengId, cleengSandbox },
   } = ConfigStore.getRawState();
-  const { user, auth } = AccountStore.getRawState();
+  const { user, auth } = useAccountStore.getState();
   const { order } = CheckoutStore.getRawState();
 
   if (!order) throw new Error('No order created');
   if (!cleengId) throw new Error('cleengId is not configured');
   if (!user || !auth) throw new Error('user is not logged in');
 
-  const response = await checkoutService.paymentWithPayPal({ orderId: order.id, successUrl, cancelUrl, errorUrl }, cleengSandbox, auth.jwt);
+  const response = await checkoutService.paymentWithPayPal(
+    {
+      orderId: order.id,
+      successUrl,
+      cancelUrl,
+      errorUrl,
+    },
+    cleengSandbox,
+    auth.jwt,
+  );
 
   if (response.errors.length > 0) throw new Error(response.errors[0]);
